rft and code cleanup

+++ homework/base_llm.py
@@ -11,7 +11,7 @@ device = "cuda" if torch.cuda.is_available() else "mps" if torch.backends.mps.is
 class BaseLLM:
     def __init__(self, checkpoint=checkpoint):
         self.tokenizer = AutoTokenizer.from_pretrained(checkpoint)
-        self.model = AutoModelForCausalLM.from_pretrained(checkpoint).to(device)
+        self.model = AutoModelForCausalLM.from_pretrained(checkpoint, torch_dtype=torch.bfloat16).to(device)
         self.device = device
 
     def format_prompt(self, question: str) -> str:
@@ -95,6 +95,7 @@ class BaseLLM:
         # Depending on your GPU batched generation will use a lot of memory.
         # If you run out of memory, try to reduce the micro_batch_size.
         micro_batch_size = 32
+        
         if len(prompts) > micro_batch_size:
             return [
                 r
